Move language code to the AdsIncentives
